chore: restructuration of folder tree
