+++ src/app/(pages)/food/generator/page.js
