+++ src/app/(pages)/food/layout.js
@@ -1,8 +1,8 @@
 import React from "react";
-import "../globals.css";
-import Navbar from "../components/food/generator/navbar";
-import { FoodModal} from "../components/food/generator/FoodModal";
-import ContextProvider from "../contextProvider";
+import "@/app/globals.css";
+import Navbar from "../../components/food/generator/navbar";
+import { FoodModal} from "../../components/food/generator/FoodModal";
+import ContextProvider from "../../contextProvider";
 
 export const metadata = {
   title: "Create Next App",
