+++ src/app/(pages)/login/layout.js
@@ -1,3 +1,5 @@
+import React from "react"
+
 export const metadata = {
   title: 'Next.js',
   description: 'Generated by Next.js',
