--- src/app/exempleAndUseCase/page.js
@@ -1,11 +1,0 @@
-import React from "react"
-
-export default function Exemples() {
-  return(
-    <div>
-      <p>this is exemple</p>
-      <button className="btn-app-yellow"> click me</button>
-      <button className="btn-app-dark"> click me</button>
-    </div>
-  )
-}
