--- src/app/food/detail/[foodId]/page.jsx
@@ -1,7 +1,0 @@
-import React from "react"
-
-export default function FoodDetail({params}){
-  return(
-    <>{params.foodId}</>
-  )
-}
